Update README.md
In the realm of data, I took on the challenge of working with the Boston Housing Dataset. My task was twofold: first, to clean up the data, ensuring it was tidy and free from inconsistencies. Then, I harnessed the power of graphs and charts to create visual representations that brought the data to life. This process transformed numbers into vivid stories, unveiling hidden insights and patterns. It was an exciting journey, and with every visualization, I uncovered more of the dataset's intriguing secrets.